requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
drf-yasg==1.20.0
python-dotenv==0.19.0
pandas==1.3.3
openpyxl==3.0.8